dm_stub : fix PRECISION.LOSS
PRECISION.LOSS
Issue:Conversion from u32 to u16 may cause loss of data.
Fix: Change size to u32.

Fixes: SYSFW-6253

+++ lpm/pll_16fft_raw.h
@@ -42,7 +42,7 @@ struct pll_raw_data {
 	u32	freq_ctrl0;
 	u32	freq_ctrl1;
 	u32	div_ctrl;
-	u16	hsdiv[16];
+	u32	hsdiv[16];
 };
 
 /**
